requirements.txt
autots==0.6.15
bayesian-optimization==2.0.0
boto3==1.35.71
botocore==1.35.71
contourpy==1.2.1
cycler==0.12.1
DateTime==5.5
folium==0.18.0
fonttools==4.53.1
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
holidays==0.61
imbalanced-learn==0.12.4
imblearn==0.0
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
joblib==1.4.2
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kagglehub==0.3.5
kiwisolver==1.4.5
lightgbm==4.5.0
lightning-utilities==0.11.8
Markdown==3.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mlflow==2.18.0
mlflow-skinny==2.18.0
nbformat==5.10.4
pytorch-lightning==1.9.5
neuralprophet==0.8.0
numpy==1.26.4
openpyxl==3.1.5
optuna==4.1.0
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pillow==10.4.0
plotly==5.23.0
pmdarima==2.0.4
psutil==6.0.0
psycopg2-binary==2.9.10
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.2
scikit-plot==0.3.7
scipy==1.14.1
scorecardpy==0.1.9.7
seaborn==0.13.2
six==1.16.0
SQLAlchemy==2.0.36
statsforecast==1.7.8
statsmodels==0.14.3
threadpoolctl==3.5.0
tzdata==2024.1
